[MOVE/#115] 기록뷰 패키징 수정
