+++ presentation/src/main/java/co/orange/presentation/history/interest/HistoryInterestAdapter.kt
@@ -1,4 +1,4 @@
-package co.orange.presentation.main.profile.history
+package co.orange.presentation.history.interest
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
